Templating cleaner, data manip in progress
